pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.2",
    "flask>=3.1.0",
    "markdown>=3.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tinydb>=4.8.2",
]
